[Android] AutocompleteMatch and AutocompleteResult are relocated
Chromium change:
https://chromium.googlesource.com/chromium/src/+/5bfb55cef9db26fdbba4ccba1a45f271f1c389fe

Relocate AutocompleteMatch and AutocompleteResult

This class aligns the naming of the OmniboxSuggestion to match its
native counterpart (AutocompleteMatch) and relocates both
AutocompleteMatch and AutocompleteResult to the
components/omnibox/browser package.

The relocation is required so that we can call the java methods
(eg. to create or update) from native code without introducing
strange build dependency.

This is part of the effort to reconcile the suggestion pools.

Bug: 1138587